Move Float8 variations to headeronly
[ghstack-poisoned]
